requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.1
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
filelock==3.13.1
flake8==7.0.0
identify==2.5.33
idna==3.6
inflection==0.5.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
nodeenv==1.8.0
oauthlib==3.2.2
platformdirs==4.1.0
pre-commit==3.6.0
psycopg==3.1.17
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
setuptools==69.0.3
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0

pytest~=7.4.4
